@@ -19,39 +19,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/cdk": "^5.0.0-rc.0",
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/http": "^5.0.2",
-    "@angular/material": "^5.0.0-rc.0",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/platform-server": "^5.0.2",
-    "@angular/router": "^5.0.2",
-    "angularjs-datepicker": "^2.1.23",
-    "bootstrap": "^4.0.0-beta.2",
+    "@angular/animations": "^5.1.0",
+    "@angular/cdk": "^5.0.1",
+    "@angular/common": "^5.1.0",
+    "@angular/compiler": "^5.1.0",
+    "@angular/core": "^5.1.0",
+    "@angular/forms": "^5.1.0",
+    "@angular/http": "^5.1.0",
+    "@angular/material": "^5.0.1",
+    "@angular/platform-browser": "^5.1.0",
+    "@angular/platform-browser-dynamic": "^5.1.0",
+    "@angular/platform-server": "^5.1.0",
+    "@angular/router": "^5.1.0",
     "core-js": "^2.5.1",
-    "font-awesome": "^4.7.0",
     "material-design-icons": "^3.0.1",
     "material-icons": "^0.1.0",
-    "popper.js": "^1.12.9",
-    "rxjs": "^5.5.2",
+    "rxjs": "5.5.2",
+    "mime-types": "^2.1.17",
+    "jszip": "^3.1.5",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.0.1",
-    "@angular/language-service": "^5.0.1",
+    "@angular/cli": "^1.6.0",
+    "@angular/compiler-cli": "^5.1.0",
+    "@angular/language-service": "^5.1.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
     "codelyzer": "~3.2.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
-    "jquery": "^3.2.1",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
@@ -64,6 +61,6 @@
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
     "tslint-eslint-rules": "^4.1.1",
-    "typescript": "2.4.2"
+    "typescript": "~2.5.3"
   }
 }
