@@ -3,8 +3,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HttpModule } from '@angular/http';
-
+import { HttpClientModule } from '@angular/common/http';
 import { NavigationComponent } from './navigation.component';
 import { CouchService } from '../shared/couchdb.service';
 import { Router } from '@angular/router';
@@ -20,7 +19,7 @@ describe('Navigation', () => {
 
   const setup = () => {
     TestBed.configureTestingModule({
-      imports: [ RouterTestingModule, CommonModule, HttpModule ],
+      imports: [ RouterTestingModule, CommonModule, HttpClientModule ],
       declarations: [ NavigationComponent ],
       providers: [
         CouchService,
