@@ -1,10 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CouchService } from '../shared/couchdb.service';
 import { MeetupsComponent } from './meetups.component';
-import { HttpModule } from '@angular/http';
 import { PlanetAlertsModule } from '../shared/alerts/planet-alerts.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
+import { HttpClientModule } from '@angular/common/http';
 import { By } from '@angular/platform-browser';
 
 describe('MeetupsComponent', () => {
@@ -22,7 +21,7 @@ describe('MeetupsComponent', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        HttpModule, PlanetAlertsModule, BrowserAnimationsModule
+         PlanetAlertsModule, BrowserAnimationsModule, HttpClientModule
       ],
       declarations: [ MeetupsComponent ],
       providers: [ CouchService ]
