@@ -1,11 +1,12 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { CouchService } from '../shared/couchdb.service';
+import { HttpClientModule } from '@angular/common/http';
 import { CourseValidatorService } from './course-validator.service';
-import { HttpModule } from '@angular/http';
+
 describe('CourseValidatorService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [ HttpModule ],
+      imports: [ HttpClientModule ],
       providers: [ CourseValidatorService, CouchService ]
     });
   });
