@@ -2,6 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'planet-app',
-  template: '<router-outlet></router-outlet>'
+  template: '<div i18n-dir dir="ltr"><router-outlet></router-outlet></div>'
 })
 export class AppComponent { }
