@@ -1,6 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { FormsModule } from '@angular/forms';
 import { CommunityComponent } from './community.component';
+import { CouchService } from '../shared/couchdb.service';
+import { AlertsDeleteComponent } from '../shared/alerts/alerts-delete.component';
+import { HttpClientModule } from '@angular/common/http';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 describe('CommunityComponent', () => {
   let component: CommunityComponent;
@@ -8,7 +12,9 @@ describe('CommunityComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CommunityComponent ]
+      imports: [ FormsModule, HttpClientModule, BrowserAnimationsModule ],
+      declarations: [ CommunityComponent, AlertsDeleteComponent ],
+      providers: [ CouchService ]
     })
     .compileComponents();
   }));
