@@ -1,6 +1,12 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { CoursesAddComponent } from './courses-add.component';
+import { FormErrorMessagesComponent } from '../../shared/form-error-messages.component';
+import { CourseValidatorService } from 'app/validators/course-validator.service';
+import { AlertsDeleteComponent } from '../../shared/alerts/alerts-delete.component';
+import { HttpClientModule } from '@angular/common/http';
+import { RouterTestingModule } from '@angular/router/testing';
+import { CouchService } from '../../shared/couchdb.service';
 
 describe('CoursesComponent', () => {
   let component: CoursesAddComponent;
@@ -8,7 +14,9 @@ describe('CoursesComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CoursesAddComponent ]
+      imports: [ ReactiveFormsModule, FormsModule, RouterTestingModule, HttpClientModule ],
+      declarations: [ CoursesAddComponent, FormErrorMessagesComponent, AlertsDeleteComponent ],
+      providers: [ CouchService, CourseValidatorService ]
     })
     .compileComponents();
   }));
