@@ -1,16 +1,20 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, AfterViewInit, ViewChild } from '@angular/core';
 import { CouchService } from '../shared/couchdb.service';
 import { DialogsDeleteComponent } from '../shared/dialogs/dialogs-delete.component';
-import { MatDialog } from '@angular/material';
+import { MatTableDataSource, MatSort, MatPaginator, MatFormField, MatFormFieldControl, MatDialog } from '@angular/material';
 
 @Component({
-  templateUrl: './courses.component.html',
-  styleUrls: [ './courses.component.scss' ]
+  templateUrl: './courses.component.html'
 })
-export class CoursesComponent implements OnInit {
+export class CoursesComponent implements OnInit, AfterViewInit {
+
+  courses = new MatTableDataSource();
+  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  displayedColumns = [ 'title', 'action' ];
   message = '';
-  courses = [];
   deleteDialog: any;
+
   constructor(
     private couchService: CouchService,
     private dialog: MatDialog
@@ -23,17 +27,29 @@ export class CoursesComponent implements OnInit {
   getCourses() {
     this.couchService.get('courses/_all_docs?include_docs=true')
       .then((data) => {
-        // don't retrieve the course validator
-        this.courses = data.rows.filter(x => x.doc._id !== '_design/course-validators');
-      }, (error) => this.message = 'There was a problem getting the courses');
+        this.courses.data = data.rows.map((course: any) => {
+          return course.doc;
+        }).filter((c: any) => {
+          return c._id !== '_design/course-validators';
+        });
+      });
+  }
+
+  ngAfterViewInit() {
+    this.courses.sort = this.sort;
+    this.courses.paginator = this.paginator;
+  }
+
+  searchFilter(filterValue: string) {
+    this.courses.filter = filterValue.trim().toLowerCase();
   }
 
   deleteClick(course) {
     this.deleteDialog = this.dialog.open(DialogsDeleteComponent, {
       data: {
-        okClick: this.deleteCourse(course.doc),
+        okClick: this.deleteCourse(course),
         type: 'course',
-        displayName: course.doc.courseTitle
+        displayName: course.courseTitle
       }
     });
   }
@@ -45,7 +61,7 @@ export class CoursesComponent implements OnInit {
       this.couchService.delete('courses/' + courseId + '?rev=' + courseRev)
         .then((data) => {
           // It's safer to remove the item from the array based on its id than to splice based on the index
-          this.courses = this.courses.filter(c => data.id !== c.doc._id);
+          this.courses.data = this.courses.data.filter((c: any) => data.id !== c._id);
           this.deleteDialog.close();
         }, (error) => this.deleteDialog.componentInstance.message = 'There was a problem deleting this course');
     };
