@@ -9,6 +9,7 @@ import { CoursesRouterModule } from './courses-router.module';
 import { PlanetFormsModule } from '../shared/planet-forms.module';
 import { CourseValidatorService } from '../validators/course-validator.service';
 import { PlanetDialogsModule } from '../shared/dialogs/planet-dialogs.module';
+import { MaterialModule } from '../shared/material.module';
 
 @NgModule({
   imports: [
@@ -17,7 +18,8 @@ import { PlanetDialogsModule } from '../shared/dialogs/planet-dialogs.module';
     FormsModule,
     ReactiveFormsModule,
     PlanetFormsModule,
-    PlanetDialogsModule
+    PlanetDialogsModule,
+    MaterialModule
   ],
   declarations: [
     CoursesComponent,
