@@ -7,6 +7,8 @@ import { HomeComponent } from './home.component';
 import { CommunityComponent } from '../community/community.component';
 import { NationComponent } from '../nation/nation.component';
 import { ManagerDashboardComponent } from '../manager-dashboard/manager-dashboard.component';
+import { RegisterComponent } from '../register/register.component';
+
 
 const routes: Routes = [
   { path: '', component: HomeComponent,
@@ -19,7 +21,8 @@ const routes: Routes = [
       { path: 'community', component: CommunityComponent },
       { path: 'community/:nation', component: CommunityComponent },
       { path: 'resources', loadChildren: '../resources/resources.module#ResourcesModule' },
-      { path: 'meetups', loadChildren: '../meetups/meetups.module#MeetupsModule' }
+      { path: 'meetups', loadChildren: '../meetups/meetups.module#MeetupsModule' },
+      { path: 'register', component: RegisterComponent }
     ]
   }
 ];
