@@ -16,6 +16,7 @@ import { NationComponent } from '../nation/nation.component';
 import { MaterialModule } from '../shared/material.module';
 import { PlanetDialogsModule } from '../shared/dialogs/planet-dialogs.module';
 import { ManagerDashboardComponent } from '../manager-dashboard/manager-dashboard.component';
+import { RegisterComponent } from '../register/register.component';
 
 @NgModule({
   imports: [
@@ -35,7 +36,8 @@ import { ManagerDashboardComponent } from '../manager-dashboard/manager-dashboar
     UsersComponent,
     CommunityComponent,
     NationComponent,
-    ManagerDashboardComponent
+    ManagerDashboardComponent,
+    RegisterComponent
   ],
   providers: [ NationValidatorService ]
 })
