@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
 import { PlanetFormsModule } from '../shared/planet-forms.module';
 import { ResourcesComponent } from './resources.component';
 import { ResourcesViewComponent } from './resources-view.component';
+import { ResourcesAddComponent } from './resources-add.component';
 import { ResourcesRouterModule } from './resources-router.module';
+import { MaterialModule } from '../shared/material.module';
+import { ResourceValidatorService } from '../validators/resource-validator.service';
 
 @NgModule({
   imports: [
@@ -13,8 +15,10 @@ import { ResourcesRouterModule } from './resources-router.module';
     FormsModule,
     ReactiveFormsModule,
     PlanetFormsModule,
-    ResourcesRouterModule
+    ResourcesRouterModule,
+    MaterialModule
   ],
-  declarations: [ ResourcesComponent, ResourcesViewComponent ]
+  declarations: [ ResourcesComponent, ResourcesViewComponent, ResourcesAddComponent ],
+  providers: [ ResourceValidatorService ]
 })
 export class ResourcesModule {}
